refactor rule name: remove "No" prefix
