+++ eslint-bridge/src/rules/main.ts
@@ -142,7 +142,7 @@ import { rule as noReferrerPolicy } from './no-referrer-policy';
 import { rule as noRequireOrDefine } from './no-require-or-define';
 import { rule as noReturnTypeAny } from './no-return-type-any';
 import { rule as noSameArgumentAssert } from './no-same-argument-assert';
-import { rule as NoSingleCharInCharacterClassesCheck } from './no-single-char-in-character-classes';
+import { rule as SingleCharInCharacterClassesCheck } from './single-char-in-character-classes';
 import { rule as noTab } from './no-tab';
 import { rule as noTryPromise } from './no-try-promise';
 import { rule as noUndefinedArgument } from './no-undefined-argument';
@@ -339,7 +339,7 @@ ruleModules['no-referrer-policy'] = noReferrerPolicy;
 ruleModules['no-require-or-define'] = noRequireOrDefine;
 ruleModules['no-return-type-any'] = noReturnTypeAny;
 ruleModules['no-same-argument-assert'] = noSameArgumentAssert;
-ruleModules['no-single-char-in-character-classes'] = NoSingleCharInCharacterClassesCheck;
+ruleModules['no-single-char-in-character-classes'] = SingleCharInCharacterClassesCheck;
 ruleModules['no-tab'] = noTab;
 ruleModules['no-try-promise'] = noTryPromise;
 ruleModules['no-undefined-argument'] = noUndefinedArgument;
