+++ eslint-bridge/src/rules/single-char-in-character-classes.ts
