+++ eslint-bridge/tests/rules/fixtures/single-char-in-character-classes.ts
