+++ eslint-bridge/tests/rules/single-char-in-character-classes.test.ts
@@ -1,1 +1,2 @@
 // TODO
+
