+++ javascript-checks/src/main/java/org/sonar/javascript/checks/CheckList.java
@@ -233,7 +233,6 @@ public final class CheckList {
       NoReturnAwaitCheck.class,
       NoReturnTypeAnyCheck.class,
       NoSameArgumentAssertCheck.class,
-      NoSingleCharInCharacterClassesCheck.class,
       NoSparseArraysCheck.class,
       NoThisAliasCheck.class,
       NoUnnecessaryTypeAssertionCheck.class,
@@ -283,6 +282,7 @@ public final class CheckList {
       SessionRegenerationCheck.class,
       ShorthandPropertiesNotGroupedCheck.class,
       SingleCharacterAlternativeCheck.class,
+      SingleCharInCharacterClassesCheck.class,
       SlowRegexCheck.class,
       SocketsCheck.class,
       SonarNoInvalidRegexCheck.class,
